test(snowflake): fold auth deprecation tests into snowflake source suite
Move the two valuable deprecation tests into test_snowflake_source.py
(next to the other SnowflakeConnectionConfig validation tests) and
delete the standalone test_snowflake_auth_deprecation.py module.

Kept:
- test_is_using_password_auth_ignores_empty_password: guards the
  bool(SecretStr("")) == True gotcha.
- test_config_validation_emits_configuration_warning_for_password_auth:
  pins the warnings.warn channel that --test-source-connection prints.

Dropped six low-value tests that pinned static string contents, obvious
short-circuiting, redundant negatives, and the add_global_warning list
append, per the repo's focus-on-value testing principles.
